speaking_club/requirements.txt
-i https://pypi.org/simple
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
crispy-bootstrap5==0.6
cryptography==40.0.2 ; python_version >= '3.6'
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==2.2.27
django-crispy-forms==1.14.0 ; python_version >= '3.7'
django-environ==0.4.4
django-jazzmin==2.6.0
django-robokassa3==1.4
gevent==22.10.2
greenlet==2.0.2 ; platform_python_implementation == 'CPython'
gunicorn==20.1.0
idna==3.4 ; python_version >= '3.5'
oauthlib==3.2.2 ; python_version >= '3.6'
psycopg2==2.9.6
pycparser==2.21
pyjwt==2.7.0 ; python_version >= '3.7'
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0 ; python_version >= '3.7'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
setuptools==67.8.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==2.1.0
https://github.com/python-social-auth/social-core/archive/master.zip
sqlparse==0.4.4 ; python_version >= '3.5'
urllib3==2.0.2 ; python_version >= '3.7'
zope.event==4.6
zope.interface==6.0 ; python_version >= '3.7'
